Extract Fortran declarator partial
